payment/requirements.txt
SQLAlchemy~=1.4.23
grpcio
grpcio-tools
stripe~=2.60.0
psycopg2-binary